+++ src/components/ScrollTop.jsx
@@ -1,0 +1,46 @@
+import React from "react";
+import { makeStyles } from "@material-ui/core/styles";
+import useScrollTrigger from "@material-ui/core/useScrollTrigger";
+import Zoom from "@material-ui/core/Zoom";
+
+const ScrollTop = (props) => {
+  const { children } = props;
+  const classes = useStyles();
+
+  const trigger = useScrollTrigger({
+    disableHysteresis: true,
+    threshold: 100,
+  });
+
+  const handleClick = (event) => {
+    const anchor = (event.target.ownerDocument || document).querySelector(
+      "#back-to-top-anchor"
+    );
+
+    if (anchor) {
+      anchor.scrollIntoView({ behavior: "smooth", block: "center" });
+    }
+  };
+
+  return (
+    <Zoom in={trigger}>
+      <div onClick={handleClick} role="presentation" className={classes.root}>
+        {children}
+      </div>
+    </Zoom>
+  );
+};
+
+const useStyles = makeStyles((theme) => ({
+  root: {
+    position: "fixed",
+    // Position bottom-right
+    bottom: theme.spacing(4),
+    right: theme.spacing(3),
+    // Position to center
+    // top: "85%",
+    // left: "47%",
+  },
+}));
+
+export default ScrollTop;
